+++ scim2_models/__init__.py
@@ -32,6 +32,7 @@ from .exceptions import SCIMException
 from .exceptions import SensitiveException
 from .exceptions import TooManyException
 from .exceptions import UniquenessException
+from .filters import ScimFilter
 from .lookup import get_model_by_payload
 from .lookup import get_model_by_schema
 from .messages.bulk import BulkOperation
@@ -152,6 +153,7 @@ __all__ = [
     "SCIMException",
     "SCIMSerializer",
     "SCIMValidator",
+    "ScimFilter",
     "Schema",
     "SchemaExtension",
     "ScimObject",